Task #6355: 판정 계약 테스트를 통합 테스트로 옮긴다
CI Lint 의 rust-unit-test-tiers 래칫이 src 신규 cfg(test) module 을 거부했다
(picture.rs::transform_invalidation_tests, 총량 4225 > 4221). 로컬 `--check` 는
PR base 비교를 하지 않아 통과했다 — `--base-ref origin/devel` 을 함께 줘야
같은 판정이 나온다.

네 테스트 모두 공개 API 만 쓰므로 tests/cases/issue_6355_picture_transform_invalidation.rs
로 옮긴다. src 쪽 cfg(test) 는 base 그대로 남는다.

- same_valued_transform_props_keep_raw_rendering
- transform_key_quoted_inside_a_value_keeps_raw_rendering
- non_transform_props_keep_raw_rendering
- changed_transform_props_still_invalidate_raw_rendering
